Alex branch

See merge request yedu/selina-server!82

+++ controllers/listingController.js
@@ -40,7 +40,10 @@ class ListingController extends Controller {
   baseCountListings = async (req, userId = null) => {
     const cities = req.body.cities ?? [];
     const categories = req.body.categories ?? [];
-    const timeInfos = await this.listTimeOption(req, 0, 2);
+    const timeInfos = await this.listTimeOption({
+      req,
+      type: STATIC.TIME_OPTIONS_TYPE_DEFAULT.TODAY,
+    });
     const searchCity = req.body.searchCity ?? null;
     const searchCategory = req.body.searchCategory ?? null;
 
@@ -143,6 +146,13 @@ class ListingController extends Controller {
   mainList = (req, res) =>
     this.baseWrapper(req, res, async () => {
       const result = await this.baseListingList(req);
+      return this.sendSuccessResponse(res, STATIC.SUCCESS.OK, null, result);
+    });
+
+  ownerList = (req, res) =>
+    this.baseWrapper(req, res, async () => {
+      const ownerId = req.body.ownerId;
+      const result = await this.baseListingList(req, ownerId);
       return this.sendSuccessResponse(res, STATIC.SUCCESS.OK, null, result);
     });
 
@@ -278,6 +288,7 @@ class ListingController extends Controller {
     dataToSave["listingImages"] = this.localGetFiles(req);
 
     const listingId = dataToSave["id"];
+    dataToSave["active"] = dataToSave["active"] == "true";
 
     const { deletedImagesInfos, listingImages: listingImagesToRes } =
       await this.listingModel.updateById(dataToSave);
@@ -306,6 +317,17 @@ class ListingController extends Controller {
       req.body["approved"] = false;
 
       const listingId = req.body.id;
+      const countUnfinishedListingOrders =
+        await this.orderModel.getUnfinishedListingCount(listingId);
+
+      if (countUnfinishedListingOrders) {
+        return this.sendErrorResponse(
+          res,
+          STATIC.ERRORS.DATA_CONFLICT,
+          "The listing has a unfinished booking or order. Please finish all listing orders and bookings before updating"
+        );
+      }
+
       const listing = await this.listingModel.getById(listingId);
 
       if (listing.approved) {
@@ -373,6 +395,18 @@ class ListingController extends Controller {
 
   baseDelete = async (req, res) => {
     const { id } = req.body;
+
+    const countUnfinishedListingOrders =
+      await this.orderModel.getUnfinishedListingCount(id);
+
+    if (countUnfinishedListingOrders) {
+      return this.sendErrorResponse(
+        res,
+        STATIC.ERRORS.DATA_CONFLICT,
+        "The listing has a unfinished booking or order. Please finish all listing orders and bookings before deleting"
+      );
+    }
+
     await this.listingApprovalRequestModel.deleteByListing(id);
     const { deletedImagesInfos } = await this.listingModel.deleteById(id);
     this.removeListingImages(deletedImagesInfos);
@@ -396,6 +430,47 @@ class ListingController extends Controller {
       );
 
       return result;
+    });
+
+  baseChangeActive = async (req, res, changeActiveCall) => {
+    const { id } = req.body;
+
+    const countUnfinishedListingOrders =
+      await this.orderModel.getUnfinishedListingCount(id);
+
+    if (countUnfinishedListingOrders) {
+      return this.sendErrorResponse(
+        res,
+        STATIC.ERRORS.DATA_CONFLICT,
+        "The listing has a unfinished booking or order. Please finish all listing orders and bookings before updating"
+      );
+    }
+
+    const active = await changeActiveCall(id);
+
+    const message = active
+      ? "Listing activated successfully"
+      : "Listing deactivated successfully";
+
+    return this.sendSuccessResponse(res, STATIC.SUCCESS.OK, message, {
+      id,
+      active,
+    });
+  };
+
+  changeActive = (req, res) =>
+    this.baseWrapper(req, res, () => {
+      const userId = req.userData.userId;
+      return this.baseChangeActive(req, res, (listingId) =>
+        this.listingModel.changeActiveByUser(listingId, userId)
+      );
+    });
+
+  changeActiveByAdmin = (req, res) =>
+    this.baseWrapper(req, res, () => {
+      return this.baseChangeActive(req, res, (listingId) =>
+        this.listingModel.changeActive(listingId)
+      );
     });
 
   deleteByAdmin = async (req, res) =>
